+++ paths/perlinNoise/src/PerlinNoiseGenerator.ts
@@ -20,6 +20,7 @@ const double = 2,
         increment: 0.004,
         columns: 0,
         rows: 0,
+        layers: 0,
         width: 0,
         height: 0,
         factor: {
@@ -29,20 +30,21 @@ const double = 2,
         offset: {
             x: 40000,
             y: 40000,
+            z: 40000,
         },
     };
 
 export class PerlinNoiseGenerator implements IMovePathGenerator {
     container?: Container;
-    field: Vector[][];
+    field: Vector[][][];
     readonly noiseGen: PerlinNoise;
-    noiseZ: number;
+    noiseW: number;
     readonly options: IPerlinOptions;
 
     constructor() {
         this.noiseGen = new PerlinNoise();
         this.field = [];
-        this.noiseZ = 0;
+        this.noiseW = 0;
         this.options = deepExtend({}, defaultOptions) as IPerlinOptions;
     }
 
@@ -52,10 +54,12 @@ export class PerlinNoiseGenerator implements IMovePathGenerator {
             point = {
                 x: Math.max(Math.floor(pos.x / size), 0),
                 y: Math.max(Math.floor(pos.y / size), 0),
+                z: Math.max(Math.floor(pos.z / size), 0),
             },
+            v = Vector.origin,
             { field } = this;
 
-        return !field?.[point.x]?.[point.y] ? Vector.origin : field[point.x][point.y].copy();
+        return field?.[point.x]?.[point.y]?.[point.z] ? field[point.x][point.y][point.z].copy() : v;
     }
 
     init(container: Container): void {
@@ -75,42 +79,43 @@ export class PerlinNoiseGenerator implements IMovePathGenerator {
 
         this._calculateField();
 
-        this.noiseZ += this.options.increment;
+        this.noiseW += this.options.increment;
 
         if (this.options.draw) {
             this.container.canvas.draw(ctx => this._drawField(ctx));
         }
     }
 
-    private readonly _calculateField: () => void = () => {
-        const { field, noiseGen, options } = this,
+    private _calculateField(): void {
+        const { field, noiseGen, options, noiseW } = this,
             lengthFactor = options.factor.length,
             angleFactor = options.factor.angle;
 
         for (let x = 0; x < options.columns; x++) {
-            const column = field[x];
-
             for (let y = 0; y < options.rows; y++) {
-                const cell = column[y];
+                for (let z = 0; z < options.layers; z++) {
+                    const cell = field[x][y][z];
 
-                cell.length = noiseGen.noise3d(
-                    x * lengthFactor + options.offset.x,
-                    y * lengthFactor + options.offset.y,
-                    this.noiseZ,
-                );
-                cell.angle = noiseGen.noise3d(x * angleFactor, y * angleFactor, this.noiseZ) * doublePI;
+                    cell.length = noiseGen.noise4d(
+                        x * lengthFactor + options.offset.x,
+                        y * lengthFactor + options.offset.y,
+                        z * lengthFactor + options.offset.z,
+                        noiseW,
+                    );
+                    cell.angle = noiseGen.noise4d(x * angleFactor, y * angleFactor, z * angleFactor, noiseW) * doublePI;
+                }
             }
         }
-    };
+    }
 
-    private readonly _drawField: (ctx: CanvasRenderingContext2D) => void = ctx => {
+    private _drawField(ctx: CanvasRenderingContext2D): void {
         const { field, options } = this;
 
         for (let x = 0; x < options.columns; x++) {
             const column = field[x];
 
             for (let y = 0; y < options.rows; y++) {
-                const cell = column[y],
+                const cell = column[y][0], // only 2D
                     { angle, length } = cell;
 
                 // ctx.save();
@@ -125,20 +130,24 @@ export class PerlinNoiseGenerator implements IMovePathGenerator {
                 // ctx.restore();
             }
         }
-    };
+    }
 
-    private readonly _initField: () => void = () => {
-        const { columns, rows } = this.options;
-        this.field = new Array<Vector[]>(columns);
+    private _initField(): void {
+        const { columns, rows, layers } = this.options;
+        this.field = new Array<Vector[][]>(columns);
 
         for (let x = 0; x < columns; x++) {
-            this.field[x] = new Array<Vector>(rows);
+            this.field[x] = new Array<Vector[]>(rows);
 
             for (let y = 0; y < rows; y++) {
-                this.field[x][y] = Vector.origin;
+                this.field[x][y] = new Array<Vector>(layers);
+
+                for (let z = 0; z < layers; z++) {
+                    this.field[x][y][z] = Vector.origin;
+                }
             }
         }
-    };
+    }
 
     private _resetField(): void {
         const container = this.container;
@@ -150,6 +159,9 @@ export class PerlinNoiseGenerator implements IMovePathGenerator {
         const sourceOptions = container.actualOptions.particles.move.path.options,
             { options } = this;
 
+        options.width = container.canvas.size.width;
+        options.height = container.canvas.size.height;
+
         options.size = (sourceOptions.size as number) > 0 ? (sourceOptions.size as number) : defaultOptions.size;
         options.increment =
             (sourceOptions.increment as number) > 0 ? (sourceOptions.increment as number) : defaultOptions.increment;
@@ -159,30 +171,29 @@ export class PerlinNoiseGenerator implements IMovePathGenerator {
 
         options.offset.x = offset?.x ?? defaultOptions.offset.x;
         options.offset.y = offset?.y ?? defaultOptions.offset.y;
+        options.offset.z = offset?.z ?? defaultOptions.offset.z;
 
         const factor = sourceOptions.factor as IFactorValues | undefined;
 
         options.factor.angle = factor?.angle ?? defaultOptions.factor.angle;
         options.factor.length = factor?.length ?? defaultOptions.factor.length;
 
-        options.width = container.canvas.size.width;
-        options.height = container.canvas.size.height;
+        options.seed = sourceOptions.seed as number | undefined;
 
-        this.options.seed = sourceOptions.seed as number | undefined;
+        this.noiseGen.seed(options.seed ?? getRandom());
 
-        this.noiseGen.seed(this.options.seed ?? getRandom());
-
-        options.columns = Math.floor(this.options.width / this.options.size) + 1;
-        options.rows = Math.floor(this.options.height / this.options.size) + 1;
+        options.columns = Math.floor(options.width / options.size) + 1;
+        options.rows = Math.floor(options.height / options.size) + 1;
+        options.layers = Math.floor(container.zLayers / options.size) + 1;
 
         this._initField();
     }
 
     private _setup(): void {
-        this.noiseZ = 0;
+        this.noiseW = 0;
 
         this._resetField();
 
-        window.addEventListener("resize", () => this._resetField());
+        addEventListener("resize", () => this._resetField());
     }
 }