+++ plugins/motion/src/MotionInstance.ts
@@ -2,7 +2,8 @@ import { type Engine, type IContainerPlugin, safeMatchMedia } from "@tsparticles
 import type { MotionContainer } from "./types.js";
 
 const defaultFactor = 1,
-    defaultReduce = 0,
+    defaultReduce = 1,
+    disableReduce = 0,
     identity = 1;
 
 export class MotionInstance implements IContainerPlugin {
@@ -67,12 +68,12 @@ export class MotionInstance implements IContainerPlugin {
 
         if (mediaQuery.matches) {
             if (motion.disable) {
-                container.retina.reduceFactor = defaultReduce;
+                container.retina.reduceFactor = disableReduce;
             } else {
                 container.retina.reduceFactor = motion.reduce.value ? identity / motion.reduce.factor : defaultFactor;
             }
         } else {
-            container.retina.reduceFactor = 1;
+            container.retina.reduceFactor = defaultReduce;
         }
     };
 }