+++ utils/configs/src/g/index.ts
@@ -2,6 +2,7 @@ import gifs from "./gifs.js";
 import grabRandomColor from "./grabRandomColor.js";
 import gradients from "./gradients.js";
 import gravity from "./gravity.js";
+import groups from "./groups";
 import growing from "./growing.js";
 
 export default {
@@ -9,5 +10,6 @@ export default {
     grabRandomColor,
     gradients,
     gravity,
+    groups,
     growing,
 };