+++ utils/configs/src/n/index.ts
@@ -2,7 +2,9 @@ import namedColors from "./namedColors.js";
 import nasa from "./nasa.js";
 import noClear from "./noClear.js";
 import noConfig from "./noConfig.js";
-import noisePlanes from "./noisePlanes.js";
+import noisePlanesFractal from "./noisePlanesFractal.js";
+import noisePlanesPerlin from "./noisePlanesPerlin.js";
+import noisePlanesSimplex from "./noisePlanesSimplex.js";
 import nyancat from "./nyancat.js";
 import nyancat2 from "./nyancat2.js";
 
@@ -11,7 +13,9 @@ export default {
     nasa,
     noClear,
     noConfig,
-    noisePlanes,
+    noisePlanesFractal,
+    noisePlanesPerlin,
+    noisePlanesSimplex,
     nyancat,
     nyancat2,
 };