+++ utils/simplexNoise/src/utils.ts
@@ -3,7 +3,7 @@
  * @param seed -
  * @returns the new seed
  */
-export function shuffleSeed(seed: Uint32Array): Uint32Array {
+export function shuffleSeed(seed: Uint32Array<ArrayBuffer>): Uint32Array<ArrayBuffer> {
     const newSeed = new Uint32Array(1);
 
     newSeed[0] = seed[0] * 1664525 + 1013904223;